+++ wdio/features/screen-objects/Modals/SkipAccountSecurityModal.js
@@ -1,0 +1,28 @@
+import {
+    SKIP_BACKUP_TEXT,
+    SKIP_BUTTON,
+  } from '../../testIDs/Screens/WalletSetupScreen.testIds';
+  import Selectors from '../../helpers/Selectors';
+  import Gestures from '../../helpers/Gestures';
+
+class SkipAccountSecurityModal {
+    
+      get skipBackupText() {
+        return Selectors.getElementByPlatform(SKIP_BACKUP_TEXT);
+      }
+      get skipButton() {
+        return Selectors.getElementByPlatform(SKIP_BUTTON);
+      }
+
+  async proceedWithoutWalletSecure(){
+    await Gestures.tap(this.skipBackupText)
+    await Gestures.tap(this.skipButton)
+  }
+
+
+  async isSkipSecurityModalVisible(){
+    await expect(this.skipBackupText).toBeDisplayed();
+  }
+
+}
+export default new SkipAccountSecurityModal();
