+++ wdio/features/screen-objects/WalletMainScreen.js
@@ -1,9 +1,14 @@
-import { ONBOARDING_WIZARD_STEP_1_CONTAINER_ID } from '../testIDs/Components/OnboardingWizard.testIds';
+import { WALLET_CONTAINER_ID } from '../testIDs/Screens/WalletScreen-testIds.js';
+import Selectors from '../helpers/Selectors';
 
 class WalletMainScreen {
-  async validateOnboardingWizard() {
-    const elem = await $(`~${ONBOARDING_WIZARD_STEP_1_CONTAINER_ID}`);
-    await expect(elem).toBeDisplayed();
+
+  get WalletScreenContainer(){
+    return Selectors.getElementByPlatform(WALLET_CONTAINER_ID)
+    
+  }
+  async isVisible() {
+    await expect(this.WalletScreenContainer).toBeDisplayed();
   }
 }
