+++ wdio/features/testIDs/Screens/WalletScreen-testIds.js
@@ -1,0 +1,1 @@
+export const WALLET_CONTAINER_ID = 'wallet-screen';
